requirements.txt
torch==2.1.2
timm==0.9.12
wandb==0.16.1
matplotlib==3.8.2
ipython==8.15.0
ipykernel==6.28.0
ipympl==0.9.3
hydra-core==1.3.2
snakeviz==2.2.0
pytorch-lightning==2.1.3
tqdm==4.65.0
torchtyping==0.1.4
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
